Move span and host device vector into public.
