+++ include/xgboost/common/span.h
