+++ include/xgboost/linear_updater.h
@@ -12,7 +12,8 @@
 #include <utility>
 #include <vector>
 #include "../../src/gbm/gblinear_model.h"
-#include "../../src/common/host_device_vector.h"
+
+#include <xgboost/common/host_device_vector.h>
 
 namespace xgboost {
 /*!
