+++ include/xgboost/metric.h
@@ -17,7 +17,7 @@
 #include <functional>
 #include <utility>
 
-#include "../../src/common/host_device_vector.h"
+#include <xgboost/common/host_device_vector.h>
 
 namespace xgboost {
 /*!
