+++ include/xgboost/objective.h
@@ -17,7 +17,7 @@
 #include <string>
 #include <functional>
 
-#include "../../src/common/host_device_vector.h"
+#include <xgboost/common/host_device_vector.h>
 
 namespace xgboost {
