+++ include/xgboost/tree_updater.h
@@ -19,7 +19,7 @@
 #include <utility>
 #include <string>
 
-#include "../../src/common/host_device_vector.h"
+#include <xgboost/common/host_device_vector.h>
 
 namespace xgboost {
 /*!
