+++ src/common/random.h
@@ -17,8 +17,8 @@
 #include <numeric>
 #include <random>
 
+#include "xgboost/common/host_device_vector.h"
 #include "io.h"
-#include "host_device_vector.h"
 
 namespace xgboost {
 namespace common {
@@ -113,7 +113,7 @@ class ColumnSampler {
   }
 
  public:
-  /** 
+  /**
    * \brief Column sampler constructor.
    * \note This constructor manually sets the rng seed
    */
@@ -169,7 +169,7 @@ class ColumnSampler {
 
   /**
    * \brief Samples a feature set.
-   * 
+   *
    * \param depth The tree depth of the node at which to sample.
    * \return The sampled feature set.
    * \note If colsample_bynode_ < 1.0, this method creates a new feature set each time it
