+++ src/gbm/gbtree.cc
@@ -6,11 +6,12 @@
  */
 #include <dmlc/omp.h>
 #include <dmlc/parameter.h>
-#include <dmlc/timer.h>
+
 #include <xgboost/logging.h>
 #include <xgboost/gbm.h>
 #include <xgboost/predictor.h>
 #include <xgboost/tree_updater.h>
+#include "xgboost/common/host_device_vector.h"
 
 #include <vector>
 #include <memory>
@@ -19,11 +20,10 @@
 #include <limits>
 #include <algorithm>
 
-#include "../common/common.h"
-#include "../common/host_device_vector.h"
-#include "../common/random.h"
 #include "gbtree.h"
 #include "gbtree_model.h"
+#include "../common/common.h"
+#include "../common/random.h"
 #include "../common/timer.h"
