+++ src/objective/regression_obj.cu
@@ -12,10 +12,11 @@
 #include <memory>
 #include <vector>
 
-#include "../common/span.h"
+#include "xgboost/common/span.h"
+#include "xgboost/common/host_device_vector.h"
+
 #include "../common/transform.h"
 #include "../common/common.h"
-#include "../common/host_device_vector.h"
 #include "./regression_loss.h"
