+++ src/predictor/cpu_predictor.cc
@@ -1,11 +1,11 @@
 /*!
  * Copyright by Contributors 2017
  */
-#include <xgboost/predictor.h>
-#include <xgboost/tree_model.h>
-#include <xgboost/tree_updater.h>
-#include "dmlc/logging.h"
-#include "../common/host_device_vector.h"
+#include "xgboost/predictor.h"
+#include "xgboost/tree_model.h"
+#include "xgboost/tree_updater.h"
+#include "xgboost/logging.h"
+#include "xgboost/common/host_device_vector.h"
 
 namespace xgboost {
 namespace predictor {
