+++ tests/cpp/common/test_span.cu
@@ -8,7 +8,7 @@
 #include <thrust/execution_policy.h>
 
 #include "../../../src/common/device_helpers.cuh"
-#include "../../../src/common/span.h"
+#include <xgboost/common/span.h>
 #include "test_span.h"
 
 namespace xgboost {
