+++ tests/cpp/tree/test_prune.cc
@@ -2,7 +2,7 @@
  * Copyright 2018-2019 by Contributors
  */
 #include "../helpers.h"
-#include "../../../src/common/host_device_vector.h"
+#include <xgboost/common/host_device_vector.h>
 #include <xgboost/tree_updater.h>
 #include <gtest/gtest.h>
 #include <vector>
